% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.3-checks.R
\name{bases_check}
\alias{bases_check}
\title{Checks for bases attribute}
\usage{
bases_check(bases)
}
\description{
Check that:
\itemize{
\item Bases are numeric vector of length 2
\item Cannot have NA values
\item Must be natural (whole) numbers greater that 0
}
}
\keyword{internal}
