% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.3-checks.R
\name{bases_equal}
\alias{bases_equal}
\title{Check that bases are equal for two deb-style vectors}
\usage{
bases_equal(x, y)
}
\description{
Used to ensure that deb_lsd and deb_decimal vectors with different bases
cannot be combined except explicitly with \code{deb_convert_bases()}.
}
\keyword{internal}
