% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-casting.R
\name{cast-decimal}
\alias{cast-decimal}
\alias{deb_as_decimal}
\alias{deb_as_decimal.deb_decimal}
\alias{deb_as_decimal.deb_lsd}
\alias{deb_as_decimal.numeric}
\alias{deb_as_decimal.logical}
\title{Cast to \code{deb_decimal}}
\usage{
deb_as_decimal(x, ...)

\method{deb_as_decimal}{deb_decimal}(x, ...)

\method{deb_as_decimal}{deb_lsd}(x, unit = c("l", "s", "d"), ...)

\method{deb_as_decimal}{numeric}(x, unit = c("l", "s", "d"), bases = c(20, 12), ...)

\method{deb_as_decimal}{logical}(x, unit = c("l", "s", "d"), bases = c(20, 12), ...)
}
\arguments{
\item{x}{An object to coerce to \code{deb_decimal}.}

\item{...}{Arguments passed on to further methods.}

\item{unit}{A character vector of length one indicating the unit for the
decimalized values, either \code{"l"} (pounds, the default), \code{"s"} (shillings),
or \code{"d"} (pence).}

\item{bases}{Numeric vector of length 2 used to specify the bases for the
solidus or s and denarius or d units. Default is \code{c(20, 12)}, which
conforms to the most widely used system of 1 pound = 20 shillings and
1 shilling = 12 pence.}
}
\description{
Cast \code{x} to a \code{deb_decimal} vector.
}
