% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casting.R
\name{deb_as_lsd}
\alias{deb_as_lsd}
\alias{deb_as_lsd.default}
\alias{deb_as_lsd.deb_lsd}
\alias{deb_as_lsd.deb_decimal}
\alias{deb_as_lsd.numeric}
\alias{deb_as_lsd.logical}
\title{Cast to deb_lsd}
\usage{
deb_as_lsd(x, ...)

\method{deb_as_lsd}{default}(x, ...)

\method{deb_as_lsd}{deb_lsd}(x, ...)

\method{deb_as_lsd}{deb_decimal}(x, ...)

\method{deb_as_lsd}{numeric}(x, bases = c(20, 12), ...)

\method{deb_as_lsd}{logical}(x, bases = c(20, 12), ...)
}
\arguments{
\item{x}{An object to coerce to \code{deb_lsd}.}

\item{...}{Arguments passed on to further methods.}

\item{bases}{Numeric vector of length 2 used to specify the bases for the
solidus or s and denarius or d units. Default is \code{c(20, 12)}, which
conforms to the most widely used system of 1 pound = 20 shillings and
1 shilling = 12 pence.}
}
\value{
A \code{deb_lsd} object.
}
\description{
Cast \code{x} to a \code{deb_lsd} vector.
}
\examples{

# Cast a deb_decimal object to deb_lsd
x <- c(5.825, 3.25, 22/3)
d1 <- deb_decimal(x)
deb_as_lsd(d1)

# Bases are automatically applied when
# casting from deb_decimal to deb_lsd
d2 <- deb_decimal(x, bases = c(60, 16))
deb_as_lsd(d2)

# Cast a numeric vector to deb_lsd
deb_as_lsd(x)

# Use the bases argument to apply non-default bases
deb_as_lsd(x, bases = c(60, 16))
}
