% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04-comparison-lsd.R
\name{comparison}
\alias{comparison}
\alias{vec_proxy_equal.deb_lsd}
\alias{vec_proxy_compare.deb_lsd}
\title{Equality and comparison}
\usage{
\method{vec_proxy_equal}{deb_lsd}(x, ...)

\method{vec_proxy_compare}{deb_lsd}(x, ...)
}
\arguments{
\item{x}{A deb_lsd object.}

\item{...}{Arguments passed on to further methods.}
}
\description{
Equality and comparison
}
