% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.2-lsd-class.R
\name{deb_bases}
\alias{deb_bases}
\title{Access the \code{bases} attribute of a \code{deb_lsd} object.}
\usage{
deb_bases(x)
}
\description{
Access the \code{bases} attribute of a \code{deb_lsd} object.
}
\keyword{internal}
