% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr-conversion.R
\name{deb_convert_unit}
\alias{deb_convert_unit}
\title{Convert the unit of \code{deb_decimal} objects}
\usage{
deb_convert_unit(x, to = c("l", "s", "d"))
}
\arguments{
\item{x}{An object of class \code{deb_decimal}.}

\item{to}{A character vector of length one indicating the unit to be
converted to. Choice of \code{"l"} (pounds, the default), \code{"s"} (shillings),
or \code{"d"} (pence).}
}
\description{
Convert the unit of \code{deb_decimal} objects
}
