% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.1-decimal-class.R
\name{deb_decimal}
\alias{deb_decimal}
\title{A decimalized class for pounds, shillings, and pence values}
\usage{
deb_decimal(x = double(), unit = c("l", "s", "d"), bases = c(20, 12))
}
\arguments{
\item{x}{A numeric vector representing the decimalized values of a
non-decimal currency.}

\item{unit}{A character vector of length one indicating the unit for the
decimalized values, either \code{"l"} (libra, the default), \code{"s"} (solidus),
or \code{"d"} (denarius).}

\item{bases}{Numeric vector of length 2 used to specify the bases for the
solidus or s and denarius or d units. Default is \code{c(20, 12)}, which
conforms to the most widely used system of 1 pound = 20 shillings and
1 shilling = 12 pence.}
}
\description{
User-facing function to create deb_decimal vector.
}
\examples{

deb_decimal(c(5.25, 3.825, 8.5))

# Set the unit of the deb_decimal vector
deb_decimal(c(105, 76.5, 170), unit = "s")
deb_decimal(c(1260, 918, 240), unit = "d")
}
