% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.1-decimal-class.R
\name{deb_is_decimal}
\alias{deb_is_decimal}
\title{Test if an object is of class \code{deb_decimal}}
\usage{
deb_is_decimal(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if object is of class \code{deb_decimal} and \code{FALSE} if it is not.
}
\description{
Test if an object is of class \code{deb_decimal}
}
