% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debvctrs-package.R
\docType{package}
\name{debvctrs-package}
\alias{debvctrs}
\alias{debvctrs-package}
\title{debvctrs: What the Package Does (One Line, Title Case)}
\description{
`debkeepr` provides an interface for analyzing non-decimal
    currencies that use the tripartite system of pounds, shillings, and pence
    through the implementation of the `lsd` class. `debkeepr` includes
    functions to apply arithmetic and financial operations to single or
    multiple values and to analyze account books that use either
    single-entry bookkeeping or double-entry bookkeeping with the latter
    providing the name for `debkeepr`.
}
\author{
\strong{Maintainer}: Jesse Sadler \email{jsadler@ucla.edu} (0000-0001-6081-9681)

}
\keyword{internal}
