% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-arithmetic-ops.R
\name{dec_arithmetic}
\alias{dec_arithmetic}
\title{Arithmetic operations for deb_decimal}
\usage{
dec_arithmetic(op, x, y)
}
\description{
Stand in for vec_arith_base() that converts unit attributes
if they are different.
}
\keyword{internal}
