% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-normalize.R
\name{decimal_check}
\alias{decimal_check}
\title{Check for non-whole numbers in libra and solidus units}
\usage{
decimal_check(lsd)
}
\description{
Move any decimals in libra and solidus units to the denarius unit.
The function uses the utility \code{should_be_int()} to deal with floating
point problems.
}
\keyword{internal}
