% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.2-lsd-class.R
\name{format.deb_lsd}
\alias{format.deb_lsd}
\title{deb_lsd format method for object printing}
\usage{
\method{format}{deb_lsd}(x, ...)
}
\description{
deb_lsd format method for object printing
}
\keyword{internal}
