% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{if_else}
\alias{if_else}
\title{vctrs ifelse function}
\usage{
if_else(test, yes, no)
}
\description{
ifelse function using vctrs to replace dplyr::if_else()
See https://vctrs.r-lib.org/articles/stability.html#ifelse
for code and explanation.
}
\keyword{internal}
