% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.3-checks.R
\name{is_natural}
\alias{is_natural}
\title{Check that bases are natural numbers}
\usage{
is_natural(x, tol = .Machine$double.eps^0.5)
}
\description{
Check that bases are natural numbers (whole number greater than 0).
From integer docs and SO: https://stackoverflow.com/a/4562291
}
\keyword{internal}
