% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-normalize.R
\name{is_negative}
\alias{is_negative}
\title{Check whether lsd value is positive or negative}
\usage{
is_negative(x)
}
\description{
Check whether lsd value is positive or negative
}
\keyword{internal}
