% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.3-checks.R
\name{lsd_check}
\alias{lsd_check}
\title{Checks for deb_lsd functions}
\usage{
lsd_check(l, s, d)
}
\description{
Checks made:
\itemize{
\item That \code{l}, \code{s}, and \code{d} are numeric
\item That they are the same length, length 1, or all length 0
}
}
\keyword{internal}
