% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-arithmetic-ops.R
\name{lsd_dividend}
\alias{lsd_dividend}
\title{Division for deb_lsd and numeric}
\usage{
lsd_dividend(x, divisor)
}
\description{
Enforce recycling rules, divide by unit, and then normalize.
}
\keyword{internal}
