% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-arithmetic-ops.R
\name{lsd_divisor}
\alias{lsd_divisor}
\title{Division for numeric over deb_lsd}
\usage{
lsd_divisor(dividend, x)
}
\description{
Enforce recycling rules, divide converted deb_decimal,
and then return to deb_lsd.
}
\keyword{internal}
