% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-arithmetic-ops.R
\name{lsd_multiply}
\alias{lsd_multiply}
\title{Multiplication for deb_lsd and numeric}
\usage{
lsd_multiply(x, multiplier)
}
\description{
Enforce recycling rules, multiply by unit, and then normalize.
}
\keyword{internal}
