% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-normalize.R
\name{lsd_normalize}
\alias{lsd_normalize}
\title{Normalization path for positive values}
\usage{
lsd_normalize(lsd)
}
\description{
Separate normalization functions for positive and negative values to be
used in \code{if_else()}. Making them functions simplifies the process.
}
\keyword{internal}
