% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper-normalize.R
\name{lsd_normalize_neg}
\alias{lsd_normalize_neg}
\title{Normalization path for negative values}
\usage{
lsd_normalize_neg(lsd)
}
\description{
Turn values positive and then return to negative value.
}
\keyword{internal}
