% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.1-decimal-class.R
\name{new_decimal}
\alias{new_decimal}
\title{Internal constructor to create \code{deb_decimal} type}
\usage{
new_decimal(x = double(), unit = c("l", "s", "d"), bases = c(20L,
  12L))
}
\description{
Asserts that \code{x} is a \code{double()}, that \code{unit} is "l", "s", or "d", and
that \code{bases} is an \code{integer()} of length 2. Creates the object through
\code{vctrs::new_vctr()}.
}
\keyword{internal}
