% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsd-class.R
\name{new_lsd}
\alias{new_lsd}
\title{Internal constructor to create deb_lsd type}
\usage{
new_lsd(l = double(), s = double(), d = double(), bases = c(20L,
  12L))
}
\value{
An object of class \code{deb_lsd}.
}
\description{
Asserts that \code{l}, \code{s} and \code{d} are of type \code{double()} and that \code{bases} is an
\code{integer()} of length 2. Creates the object through \code{vctrs::new_rcrd()}.
}
\keyword{internal}
