% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.1-decimal-class.R
\name{obj_print_footer.deb_decimal}
\alias{obj_print_footer.deb_decimal}
\title{Print footer with unit and bases}
\usage{
\method{obj_print_footer}{deb_decimal}(x, ...)
}
\description{
Print footer with unit and bases
}
\keyword{internal}
