% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.2-lsd-class.R
\name{obj_print_footer.deb_lsd}
\alias{obj_print_footer.deb_lsd}
\title{Print footer with bases}
\usage{
\method{obj_print_footer}{deb_lsd}(x, ...)
}
\description{
Print footer with bases
}
\keyword{internal}
