% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.1-decimal-class.R
\name{unit_word}
\alias{unit_word}
\title{Print full name of unit in footer}
\usage{
unit_word(x)
}
\arguments{
\item{x}{A deb_decimal object.}
}
\description{
Print full name of unit in footer
}
\keyword{internal}
