% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-coercion.R, R/03-casting.R,
%   R/04-comparison-lsd.R, R/05-mathematical-funcs.R, R/06-arithmetic-ops.R,
%   R/debvctrs-package.R
\name{vec_ptype2.deb_decimal}
\alias{vec_ptype2.deb_decimal}
\alias{vec_ptype2.deb_lsd}
\alias{vec_cast.deb_decimal}
\alias{vec_cast.deb_lsd}
\alias{vec_proxy_equal.deb_lsd}
\alias{vec_proxy_compare.deb_lsd}
\alias{vec_math.deb_lsd}
\alias{vec_arith.deb_decimal}
\alias{vec_arith.deb_decimal.default}
\alias{vec_arith.deb_decimal.deb_decimal}
\alias{vec_arith.deb_decimal.numeric}
\alias{vec_arith.numeric.deb_decimal}
\alias{vec_arith.deb_decimal.MISSING}
\alias{vec_arith.deb_lsd}
\alias{vec_arith.deb_lsd.default}
\alias{vec_arith.deb_lsd.deb_lsd}
\alias{vec_arith.deb_lsd.numeric}
\alias{vec_arith.numeric.deb_lsd}
\alias{vec_arith.deb_lsd.MISSING}
\alias{vec_arith.deb_lsd.deb_decimal}
\alias{vec_arith.deb_decimal.deb_lsd}
\alias{vctrs-compat}
\title{vctrs compatibility functions for math, casting, coercion, and comparison}
\usage{
\method{vec_ptype2}{deb_decimal}(x, y, ...)

\method{vec_ptype2}{deb_lsd}(x, y, ...)

\method{vec_cast}{deb_decimal}(x, to, ...)

\method{vec_cast}{deb_lsd}(x, to, ...)

\method{vec_proxy_equal}{deb_lsd}(x, ...)

\method{vec_proxy_compare}{deb_lsd}(x, ...)

\method{vec_math}{deb_lsd}(.fn, .x, ...)

\method{vec_arith}{deb_decimal}(op, x, y)

\method{vec_arith.deb_decimal}{default}(op, x, y)

\method{vec_arith.deb_decimal}{deb_decimal}(op, x, y)

\method{vec_arith.deb_decimal}{numeric}(op, x, y)

\method{vec_arith.numeric}{deb_decimal}(op, x, y)

\method{vec_arith.deb_decimal}{MISSING}(op, x, y)

\method{vec_arith}{deb_lsd}(op, x, y)

\method{vec_arith.deb_lsd}{default}(op, x, y)

\method{vec_arith.deb_lsd}{deb_lsd}(op, x, y)

\method{vec_arith.deb_lsd}{numeric}(op, x, y)

\method{vec_arith.numeric}{deb_lsd}(op, x, y)

\method{vec_arith.deb_lsd}{MISSING}(op, x, y)

\method{vec_arith.deb_lsd}{deb_decimal}(op, x, y)

\method{vec_arith.deb_decimal}{deb_lsd}(op, x, y)
}
\arguments{
\item{x, y}{Vectors.}

\item{...}{Arguments passed on to further methods.}

\item{to}{Type to cast to.}

\item{.fn}{Mathematical function.}

\item{.x}{Vector.}

\item{op}{Arithmetic operation.}
}
\description{
These functions are the extensions that allow debkeepr vectors to
work with vctrs.
}
