% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.1-decimal-class.R
\name{vec_ptype_abbr.deb_decimal}
\alias{vec_ptype_abbr.deb_decimal}
\title{Abbreviated name for tibble columns}
\usage{
\method{vec_ptype_abbr}{deb_decimal}(x)
}
\description{
Abbreviated name for tibble columns
}
\keyword{internal}
