% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.2-lsd-class.R
\name{vec_ptype_abbr.deb_lsd}
\alias{vec_ptype_abbr.deb_lsd}
\title{Abbreviated name for tibble columns}
\usage{
\method{vec_ptype_abbr}{deb_lsd}(x)
}
\description{
Abbreviated name for tibble columns
}
\keyword{internal}
