% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-arithmetic-ops.R
\name{arithmetic}
\alias{arithmetic}
\alias{vec_arith.deb_decimal}
\alias{vec_arith.deb_decimal.default}
\alias{vec_arith.deb_decimal.deb_decimal}
\alias{vec_arith.deb_decimal.numeric}
\alias{vec_arith.numeric.deb_decimal}
\alias{vec_arith.deb_decimal.MISSING}
\alias{vec_arith.deb_lsd}
\alias{vec_arith.deb_lsd.default}
\alias{vec_arith.deb_lsd.deb_lsd}
\alias{vec_arith.deb_lsd.numeric}
\alias{vec_arith.numeric.deb_lsd}
\alias{vec_arith.deb_lsd.MISSING}
\alias{vec_arith.deb_lsd.deb_decimal}
\alias{vec_arith.deb_decimal.deb_lsd}
\title{Arithmetic operations for debvctrs}
\usage{
\method{vec_arith}{deb_decimal}(op, x, y)

\method{vec_arith.deb_decimal}{default}(op, x, y)

\method{vec_arith.deb_decimal}{deb_decimal}(op, x, y)

\method{vec_arith.deb_decimal}{numeric}(op, x, y)

\method{vec_arith.numeric}{deb_decimal}(op, x, y)

\method{vec_arith.deb_decimal}{MISSING}(op, x, y)

\method{vec_arith}{deb_lsd}(op, x, y)

\method{vec_arith.deb_lsd}{default}(op, x, y)

\method{vec_arith.deb_lsd}{deb_lsd}(op, x, y)

\method{vec_arith.deb_lsd}{numeric}(op, x, y)

\method{vec_arith.numeric}{deb_lsd}(op, x, y)

\method{vec_arith.deb_lsd}{MISSING}(op, x, y)

\method{vec_arith.deb_lsd}{deb_decimal}(op, x, y)

\method{vec_arith.deb_decimal}{deb_lsd}(op, x, y)
}
\arguments{
\item{op}{Arithmetic operation.}

\item{x, y}{Vectors.}
}
\description{
Arithmetic operations for debvctrs
}
