% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attr-conversion.R
\name{convert-bases}
\alias{convert-bases}
\alias{deb_convert_bases}
\alias{deb_convert_bases.deb_lsd}
\alias{deb_convert_bases.deb_decimal}
\title{Convert bases of \code{deb_lsd} and \code{deb_decimal} objects}
\usage{
deb_convert_bases(x, to)

\method{deb_convert_bases}{deb_lsd}(x, to)

\method{deb_convert_bases}{deb_decimal}(x, to)
}
\arguments{
\item{x}{An object of class \code{deb_lsd} or \code{deb_decimal}.}

\item{to}{Numeric vector of length 2, representing the bases for the
solidus and denarius units to be converted to.}
}
\description{
\code{deb_convert_bases()} is the only way to change the bases of the solidus
and denarius units associated with objects of class \code{deb_lsd} or \code{deb_lsd}.
}
