% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversion.R
\name{deb_convert_unit}
\alias{deb_convert_unit}
\title{Convert the unit of a \code{deb_decimal} object}
\usage{
deb_convert_unit(x, to = c("l", "s", "d"))
}
\arguments{
\item{x}{An object of class \code{deb_decimal}.}

\item{to}{A character vector of length one indicating the unit to be
converted to. Choice of \code{"l"} (libra, the default), \code{"s"} (solidus),
or \code{"d"} (denarius).}
}
\value{
A \code{deb_decimal} object with a converted unit.
}
\description{
Convert the \code{unit} attribute of a \code{deb_decimal} object.
}
\details{
\code{deb_convert_unit()} converts the \code{unit} of a \code{deb_decimal} object to
either \code{"l"}, \code{"s"}, or \code{"d"}. This changes the representation of the
object, but the value remains equivalent.
}
\examples{

x <- deb_decimal(c(8.825, 15.125, 3.65))
y <- deb_decimal(c(56.45, 106.525, 200.4), unit = "s")

deb_convert_unit(x, to = "s")
deb_convert_unit(x, to = "d")
deb_convert_unit(y, to = "l")
deb_convert_unit(y, to = "d")
}
