% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decimal-class.R
\name{deb_is_decimal}
\alias{deb_is_decimal}
\title{Test if an object is of class \code{deb_decimal}}
\usage{
deb_is_decimal(x)
}
\arguments{
\item{x}{An object.}
}
\value{
\code{TRUE} if object is of class \code{deb_decimal} and \code{FALSE} if it is not.
}
\description{
Test if an object is of class \code{deb_decimal}.
}
\examples{
x <- deb_decimal(c(5.25, 3.825, 8.5))
y <- c(5.25, 3.825, 8.5)

deb_is_lsd(x)
deb_is_lsd(y)
}
