% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01.2-lsd-class.R
\name{deb_lsd}
\alias{deb_lsd}
\title{A class for pounds, shillings and pence values}
\usage{
deb_lsd(l = double(), s = double(), d = double(), bases = c(20, 12))
}
\arguments{
\item{l}{Numeric vector representing the pounds unit.}

\item{s}{Numeric vector representing the shillings unit.}

\item{d}{Numeric vector representing the pence unit.}

\item{bases}{Numeric vector of length 2 used to specify the bases for the
solidus or s and denarius or d units. Default is \code{c(20, 12)}, which
conforms to the most widely used system of 1 pound = 20 shillings and
1 shilling = 12 pence.}
}
\description{
User-facing function to create deb_lsd vector.
}
