% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debvctrs-package.R
\docType{package}
\name{debvctrs-package}
\alias{debvctrs}
\alias{debvctrs-package}
\title{debvctrs: Tutorial for building S3 vectors with vctrs}
\description{
Based on the more complete debkeepr package, this package
    is a tutorial to help demonstrate the process for creating custom
    vector classes with the vctrs package.
}
\author{
\strong{Maintainer}: Jesse Sadler \email{jsadler@ucla.edu} (\href{https://orcid.org/0000-0001-6081-9681}{ORCID})

}
\keyword{internal}
