% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/casting.R, R/coercion.R, R/comparison.R,
%   R/debvctrs-package.R
\name{vec_cast.deb_lsd}
\alias{vec_cast.deb_lsd}
\alias{vec_cast.deb_decimal}
\alias{vec_type2.deb_lsd}
\alias{vec_type2.deb_decimal}
\alias{vec_proxy_equal.deb_lsd}
\alias{vec_proxy_compare.deb_lsd}
\alias{vctrs-compat}
\title{vctrs compatibility functions for casting, coercion, and comparison}
\usage{
\method{vec_cast}{deb_lsd}(x, to)

\method{vec_cast}{deb_decimal}(x, to)

\method{vec_type2}{deb_lsd}(x, y)

\method{vec_type2}{deb_decimal}(x, y)

\method{vec_proxy_equal}{deb_lsd}(x)

\method{vec_proxy_compare}{deb_lsd}(x)
}
\arguments{
\item{x, y}{Vectors.}

\item{to}{Type to cast to.}
}
\description{
These functions are the extensions that allow debkeepr vectors to
work with vctrs.
}
